+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add level locking for individual buildings.
+
 ### 0.2.0.1
 - Downgrade Harmony to 2.2.2 to keep TDW happy.